CMS: Improve Content File and HTML Change Detection.
FossilOrigin-Name: 2c8c47ffa98397032634197814c6f81712018cbf2b020a489c919eef3bd7c522

+++ PointlessWaymarks.CmsData/ContentHtml/GeoJsonHtml/SingleGeoJsonPageCode.cs
@@ -65,6 +65,9 @@ public partial class SingleGeoJsonPage
 
         if (htmlFileInfo.Exists)
         {
+            if (await FileManagement.FileAndHtmlStringIgnoringContentAndGenerationVersionAreEqual(htmlFileInfo, htmlString))
+                return;
+
             htmlFileInfo.Delete();
             htmlFileInfo.Refresh();
         }
