+++ PointlessWaymarks.CmsData/ContentHtml/IndexHtml/IndexPageCode.cs
@@ -235,6 +235,9 @@ public partial class IndexPage
 
         if (htmlFileInfo.Exists)
         {
+            if (await FileManagement.FileAndHtmlStringIgnoringContentAndGenerationVersionAreEqual(htmlFileInfo, htmlString))
+                return;
+            
             htmlFileInfo.Delete();
             htmlFileInfo.Refresh();
         }
