+++ PointlessWaymarks.CmsData/ContentHtml/PhotoGalleryHtml/CameraRollGalleryPageCode.cs
@@ -24,6 +24,9 @@ public partial class CameraRollGalleryPage
 
         if (htmlFileInfo.Exists)
         {
+            if (await FileManagement.FileAndHtmlStringIgnoringContentAndGenerationVersionAreEqual(htmlFileInfo, htmlString))
+                return;
+
             htmlFileInfo.Delete();
             htmlFileInfo.Refresh();
         }
